pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gato-x"
version = "0.5"
description = "GitHub Actions Enumeration and Attack Framework"
readme = "README.md"
authors = [
    {name = "Adnan Khan", email = "adnanthekhan@proton.me"}
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "colorama",
    "requests",
    "pyyaml",
    "packaging",
    "cryptography"
]

[project.optional-dependencies]
test = [
    "flake8",
    "pytest",
    "pytest-cov"
]

[tool.hatch.build.targets.wheel]
packages = [
    "gatox"
]

[tool.pytest.ini_options]
testpaths = ["unit_test"]
python_files = ["test_*.py"]
addopts = "--cov=gatox"

[tool.coverage.run]
source = ["gatox"]

[tool.coverage.html]
directory = "cov_html"

[project.scripts]
gato-x = "gatox.main:entry"
